move mcstate and mcmixedstate to their own modules
fix imports and define modules


move again


fixes


move mcstate and mcmixedstate to their own modules


fix imports and define modules


test


impro


a


yo


simjplify


a


simpler


fixup


a


format
