+++ netket/vqs/mc/__init__.py
@@ -12,5 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from .common import check_hilbert, get_configs, get_fun
+
 from .mc_state import MCState
 from .mc_mixed_state import MCMixedState
