+++ netket/vqs/mc/mc_mixed_state/__init__.py
@@ -13,3 +13,7 @@
 # limitations under the License.
 
 from .state import MCMixedState
+
+from . import expect
+
+from . import expect_grad
